+++ src/api/team.ts
@@ -1,0 +1,7 @@
+import { BaseResponse, TeamResponse } from '@/types';
+import http from '@/api/core';
+
+export const getTeams = (): Promise<BaseResponse<TeamResponse>> =>
+  http.get({
+    url: '/teams',
+  });
