+++ src/types/dto/index.ts
@@ -1,3 +1,4 @@
 export * from './adminMember';
 export * from './base';
 export * from './applicationForm';
+export * from './team';
